refactor: account 패키지 리팩토링
