+++ src/main/java/com/mobble/mobbleserver/account/auth/dto/request/SocialLoginRequestDto.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.account.auth.dto.request;
 
-import com.mobble.mobbleserver.account.oauth2.service.SocialProvider;
+import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
 
 public record SocialLoginRequestDto(
         String accessToken,
