+++ src/main/java/com/mobble/mobbleserver/account/auth/dto/response/SignUpDetailsInfoResponseDto.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.account.auth.dto.response;
 
-import com.mobble.mobbleserver.account.oauth2.service.SocialProvider;
-import com.mobble.mobbleserver.account.oauth2.verifier.SocialUserInfo;
+import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
+import com.mobble.mobbleserver.account.auth.oauth.verifier.SocialUserInfo;
 
 public record SignUpDetailsInfoResponseDto(
         String name,
