+++ src/main/java/com/mobble/mobbleserver/account/auth/oauth/verifier/AbstractSocialTokenVerifier.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.account.oauth2.verifier;
+package com.mobble.mobbleserver.account.auth.oauth.verifier;
 
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.oAuth2.OAuth2ErrorCode;
