+++ src/main/java/com/mobble/mobbleserver/account/auth/oauth/verifier/SocialUserInfo.java
@@ -1,6 +1,6 @@
-package com.mobble.mobbleserver.account.oauth2.verifier;
+package com.mobble.mobbleserver.account.auth.oauth.verifier;
 
-import com.mobble.mobbleserver.account.oauth2.service.SocialProvider;
+import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
 
 public record SocialUserInfo(
         String name,
