+++ src/main/java/com/mobble/mobbleserver/account/auth/oauth/verifier/SocialVerifier.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.account.oauth2.verifier;
+package com.mobble.mobbleserver.account.auth.oauth.verifier;
 
 public interface SocialVerifier {
     SocialUserInfo verify(String accessToken);
