+++ src/main/java/com/mobble/mobbleserver/domain/member/entity/Member.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.domain.member.entity;
 
-import com.mobble.mobbleserver.account.oauth2.service.SocialProvider;
+import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
 import com.mobble.mobbleserver.common.baseEntity.BaseEntity;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.member.MemberErrorCode;
